ReleaseNotes.xml: Update credits section for 7.6 RC1
The list of people to credit in this release was created by running
the script at https://gist.github.com/671819 with my current list of
katamari modules/versions to get a unified ChangeLog for all katamari
modules.   For xcb modules, since this is their first inclusion, this
went back to the dawn of the xcb project.

From that list, everyone listed in the long form log output under either
Author: or *-by: was collected and added to the "These people contributed
in some way to X11R&relvers; since the release of X11R&prevrelvers;" list.

The output was manually edited to remove duplicates which appeared to be
the same person, and non-identifiable usernames ("root", "xgi0007", etc.)
Additionally, the lead freedesktop.org sysadmin was manually added.

For the second list, "This product includes software developed by",
the COPYING files in all katamari modules were mined for Copyright
notices and merged with the existing list.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>