Solaris: shorten the title a bit
OpenSolaris has been deprecated in favor of Solaris Express
The Solaris brand name is generic enough for all flavors of Solaris
X.Org is not mentioned in the title of any document.
Now that "X Version 11, Release 7.6" has been added, it should make
the title page as informative as any other document.

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>